Disconnect/Move edge
Edges can now be pulled to disconnect them from a connector, allowing
them to be moved or deleted as with creating a new edge from a connector.

Updated removeEdge and detecting a duplicate history in model/Graph
so that they work correctly with immutable Lists. Moving edges
works as expected with history undo/redo.