Connector colour based on direction and updated edge moving
Clicking on an input connector no longer creates a new edge. If the
connector has an edge then that edge can be moved.
Clicking on an output connector creates a new edge from that connector
if maxConnections hasn't been reached.
'Pulling' an edge works as before, disconnecting it from its closest
connector.