Complete overhaul of subgraph creating and editing
Processed and raw subgraph, clone and selector graph data is no
longer created and stored for use in creating the UI. Instead
the subgraph/clone/selector data is requested on entry. This keeps
the UI up to date (e.g. when adding a new node to a subgraph, exiting
and re-entering) without the need to manipulate the stored data.

Also improved the functionality of the graph selector (see Bug#1482).

BUG:FIXED:1412