Subgraph Pins
Subgraph pin nodes can now be created using the menu as normal, with
created nodes having different css. When a pin is connected to another node
the appropriate direction is sent to the parent node and the pin
becomes an input/output pin. If the connection is removed then the
input/output is removed from the parent.
On 'back' that graph is regenerated (from rest) so that it is up
to date (input/output connectors on subgraph depending on pins). Will
probably need to optimise this in the future.
Note: Undo/Redo doesn't change subgraph input/output at the moment.

Added category to InfoPanel
Generalised Menu (const names)
Non property post and delete added to vgData (for pins)
Couple of general improvements to back and back button hide/show

BUG:COMMENT:1537